Move run command implementation to pipeline plugin
