+++ python/michelangelo/cli/mactl/plugins/pipeline/main.py
@@ -6,6 +6,8 @@ from mactl import CRD
 from plugins.pipeline.apply import generate_apply
 from plugins.pipeline.create import generate_create
 from plugins.pipeline.run import generate_run
+from plugins.pipeline.run import convert_crd_metadata_pipeline_run, generate_run
+from mactl import CRD
 
 
 _LOG = getLogger(__name__)
@@ -22,8 +24,36 @@ def apply_plugins(
         generate_apply(crd, channel)
     if target_command == "create":
         generate_create(crd, channel)
-    if target_command == "run":
-        from plugins.pipeline.run import convert_crd_metadata_pipeline_run
-        crd.func_crd_metadata_converter = convert_crd_metadata_pipeline_run
-        generate_run(crd, channel)
     _LOG.info("Plugins applied successfully to crd: %s", crd)
+
+
+def handle_pipeline_command(args: list[str], kwargs: dict[str, str], channel: Channel):
+    """
+    Handle pipeline-specific commands like 'run'.
+    
+    Args:
+        args: Command arguments (e.g., ["pipeline", "run"])
+        kwargs: Command options (e.g., {"namespace": "test", "name": "pipeline"})
+        channel: gRPC channel
+    
+    Returns:
+        Response message or None if command not handled
+    """
+    if len(args) < 2:
+        return None
+        
+    resource_type = args[0]  
+    action = args[1]         
+    
+    if resource_type != "pipeline":
+        return None
+        
+    if action == "run":
+        _LOG.info("Handling pipeline run command")
+        pipeline_crd = CRD(name="pipeline", full_name="michelangelo.api.v2.PipelineService")
+        pipeline_crd.func_crd_metadata_converter = convert_crd_metadata_pipeline_run
+        
+        generate_run(pipeline_crd, channel)
+        return pipeline_crd.run(**kwargs)
+    
+    return None
